requirements.txt
Flask==2.3.3
flask-cors==4.1.1
PyPDF2==3.2.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.2.0
groq==1.1.2